Extract src/tray.ts — the native menu-bar mirror
Third of the renderAll() hot-path surfaces. Moved whole (one guarded line
range): lastTraySig and updateTray.

Unlike every other render surface this one is native, so a repaint is an IPC
call and a menu rebuild rather than an innerHTML assignment — which is why
it already diffs a signature of what it would draw and returns early when
nothing changed. Isolating it is what lets the Phase-3 profiling pass ask
whether that guard is enough, since renderAll() calls it on every telemetry
event.

Needed no hooks at all: with the needs-you cluster now in ./grouping every
callee is already below it.

The only non-`export` difference from HEAD is a deleted comment — the
`// ---------- macOS menu-bar (tray) mirror of the sidebar ----------`
section divider, which read as noise directly under the module docblock
that now says the same thing. No code changed.

No tests: render module, per the Phase-1 rule. main.ts 2,280 -> 2,234.
Gates: tsc --noEmit clean, 368 vitest pass. No Rust touched.
